Restructure resource
