+++ src/client/app/teams/tags/index.ts
@@ -2,5 +2,7 @@ export * from './list-tags.component';
 export * from './manage-tag.component';
 export * from './audit/tag-audit.component';
 
+export * from './tags.service';
+
 export * from './tags.module';
 export * from './tags-routing.module';
